♻️ refactor: core presigned_url 서비스 리팩터링
- command.py --> presigned_url.py
- Enum 클래스 정의 개별 앱의 뷰로 이동
